+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support required fields for Avro schema export (#390)
 - Support for import of DBML Models (#379)
 - `datacontract export --format sqlalchemy`: Export to [SQLAlchemy ORM models](https://docs.sqlalchemy.org/en/20/orm/quickstart.html) (#399)
+- Support data type map in Spark import and export (#408)
 
 ### Changed
 - Deltalake is now using DuckDB's native deltalake support (#258). Extra deltalake removed.
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Corrected name of table parameter for bigquery import (#377)
 - Fix an failed to connect to S3 Server (#384)
 - Fix a model bug mismatching with the specification (`definitions.fields`) (#375)
+- Fix array type management in Spark import (#408)
 
 
 ## [0.10.11] - 2024-08-08